% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_to_decimal.R
\name{base_to_decimal10}
\alias{base_to_decimal10}
\title{Find the decimal representation of any numerical system in any base.}
\usage{
base_to_decimal10(digits, base)
}
\arguments{
\item{digits}{A digit vector in a specific base}

\item{base}{The base of your digits}
}
\value{
A numeric object, The decimal (base 10) answer.
}
\description{
Solve any numerical system in a given base easily.
}
\examples{
base_to_decimal10(digits=10010,base=2)
}
