% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2b.R
\name{beta2b}
\alias{beta2b}
\title{Convert Beta to b}
\usage{
beta2b(beta, x, y)
}
\arguments{
\item{beta}{The beta value you want to convert.}

\item{x}{The x variable (Undependent).}

\item{y}{The y variable (Dependent).}
}
\value{
raw Mekadem b
}
\description{
Converts the Beta value to a raw b, given the x and y
}
\examples{
model = lm(Bsum~sat+age,data=depression)
by using \code{\link{lm.beta}} we get sat~0.504, and age~0.467.
We want to convert them back to raw b:
sat: beta2b(beta=0.5038211, x=depression$sat, y=depression$Bsum)
## 0.7899105
age: beta2b(0.4673743,depression$age, y=depression$Bsum)
## 0.4072746
}
