% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big5.R
\docType{data}
\name{big5}
\alias{big5}
\title{Big 5 personality traits}
\format{
A data frame with 250 rows and 7 variables:
\describe{
  \item{OCEAN}{Score in each of the 5 personality traits}
  \item{age}{Individual's age}
  \item{gender}{Individual's gender}
}
}
\usage{
data(big5)
}
\description{
A dataset containing the score in each of the personality traits, age and gender for each observation.
}
\details{
\url{https://www.psychologytoday.com/intl/basics/big-5-personality-traits}
}
\references{
BGU, Psychology, Regression
}
\keyword{datasets}
