% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bikoret.finder.R
\name{BikoretFinder}
\alias{BikoretFinder}
\title{ID Bikoret number generator}
\usage{
BikoretFinder(id)
}
\arguments{
\item{id}{Only input is your id without bikoret.
#' @examples BikoretFinder("12345678")
# returns 2}
}
\description{
Not a very useful command :)
use "quotes" around your ID no.
}
