% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.R
\name{ci}
\alias{ci}
\title{Confidence Interval for a known sd in population}
\usage{
ci(avg, sd, N, conf = 95)
}
\arguments{
\item{avg}{avg found in sample.}

\item{sd}{SD in the population.}

\item{N}{sample size.}

\item{conf}{confidence level. default is 95.}
}
\description{
Use this function when SD in population is known.
}
\examples{
ci(620,15,100)
# is 620 the mean in population, 95\% confidence
ci(avg=100,sd=20,N=50,conf=99)
# confidence is now at 99\%
}
\seealso{
\code{\link{ci.t}} if SD in population is unknown.
}
