% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci.t.R
\name{ci.t}
\alias{ci.t}
\title{Confidence Interval using a t distribution}
\usage{
ci.t(avg, sd, N, conf = 95)
}
\arguments{
\item{avg}{Avg found in sample}

\item{sd}{SD in the sample}

\item{N}{Sample size}

\item{conf}{Confidence level for CI. default 95}
}
\description{
Use this function when SD in population is unknown.
}
\examples{
\code{\link{ci.t(avg=25,sd=12,N=30)}}
\code{\link{ci.t(avg=25,sd=12,N=30,confidence=99)}}
\code{\link{ci.t(620,15,100)}}
# is 620 the mean in population, 95\% confidence
ci.t(avg=100,sd=20,N=50,conf=99)
# confidence is now at 99\%
}
\seealso{
\code{\link{ci()}} if SD in population is known
}
