% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohend.param.R
\name{cohend.param}
\alias{cohend.param}
\title{Cohen's d effect size for two groups}
\usage{
cohend.param(mu1, mu2, s1, s2, N1, N2 = N1, h0 = 0)
}
\arguments{
\item{mu1}{Avg of 1st group.}

\item{mu2}{Avg of 2nd group.}

\item{s1}{Variance of 1st group (NOT SD).}

\item{s2}{Variance of 2nd group (NOT SD).}

\item{N1}{Sample size of 1st group.}

\item{N2}{Sample size of 2nd group. default is N2=N1.}

\item{h0}{Null hypothesis. default is 0.}
}
\description{
Calculates effect size (Cohen's d) with given parameters.
}
\examples{
\code{\link{cohend.param(mu1=835,mu2=831,s1=8.5,s2=12,N1=11)}}
}
\seealso{
If you have vectors, use \code{\link{cohend.vec}}
}
