% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohend.R
\name{cohend}
\alias{cohend}
\title{Cohen's d effect size for one or two groups.}
\source{
\url{https://www.rdocumentation.org/packages/effsize/versions/0.8.1/topics/cohen.d}
}
\usage{
cohend(x, y, conf = 95, mu = 0, g = FALSE)
}
\arguments{
\item{x}{First vector.}

\item{y}{Second vector.}

\item{conf}{Desired confidence level. default is set to 95.}

\item{mu}{State your prior belief (H0) for the true difference between the groups. defaults is set to 0.}

\item{g}{calls for the 'Hedge's g' correction. default is set to FALSE.}
}
\description{
Calculates effect size (cohen's d) for one or two samples.
}
\details{
Full details via the \code{\link{effsize}} documantation. You can use either two vectors \code{x} and \code{y},
Or set \code{x} as a vector and \code{y} as a factor variable
}
\examples{
x <- big5$Agreeableness[big5$gender=="Females"]
y <- big5$Agreeableness[big5$gender=="Males"]
cohend(x,y)
# We comapre Agreeableness level between both gender, using two seperate vectors.
cohend(x=big5[,1],y=big5$gender)
# We comapre the same variables, now y is a factoring varaible. Also used '[]' for shorter script.
cohend(big5$Agreeableness,y=NA,mu=1,conf=99)
# Only one vector compparison against Null hypothesis that mu equals 1, in 99\% confidence level.
## Note that you need to set y=NA if you dont want to use it, otherwise you'll get an Error.
}
\seealso{
This function makes use of \code{\link{cohen.d}} from \code{\link{effsize}} package.
}
