% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohend.vec.R
\name{cohend.vec}
\alias{cohend.vec}
\title{Cohen's d effect size for two groups}
\usage{
cohend.vec(x, y, h0 = 0)
}
\arguments{
\item{x}{First vector.}

\item{y}{Second vector.}

\item{h0}{Null hypothesis. default is 0}
}
\description{
Calculates effect size (Cohen's d) with given vectors.
}
\examples{
x <- c(14,17,19,22)
y <- c(15,19,20,12)
\code{\link{cohend.vec(x,y)}}
# in this example, Null hypothesis is that reall diff is 0
\code{\link{cohend.vec(x,y,h0=2)}}
# Here, diff in pop (H0) is 2
}
\seealso{
If you have parameters, use \code{\link{cohend.param}}.
}
