% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deg.to.rad.R
\name{deg.to.rad}
\alias{deg.to.rad}
\title{Degrees to Radians}
\usage{
deg.to.rad(deg)
}
\description{
transforms degrees to radians
}
\examples{
deg.to.rad(360)
# returns 6.283185, which indeed equals to '2*pi'
}
