% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression.R
\docType{data}
\name{depression}
\alias{depression}
\title{Depression table with predicting variables.}
\format{
A data frame with 433 rows and 6 variables:
\describe{
  \item{Bsum}{Depression score}
  \item{self_income}{How much does the individual earns, in US dollars per year}
  \item{partner_income}{How much does the individual's partner earns, in US dollars per year}
  \item{age}{The individual's age}
  \item{sat}{Satisfaction score}
  \item{m_years}{Years married}
}
}
\source{
\url{https://in.bgu.ac.il/humsos/psych/eng/Pages/ContactUs.aspx}
}
\usage{
data(depression)
}
\description{
A dataset containing the depression score and other variables of 433 people.
}
\references{
BGU, Psychology, Regression
}
\keyword{datasets}
