% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.binom}
\alias{draw.binom}
\title{Draws a binomal distribution}
\usage{
draw.binom(
  X,
  n = max(X),
  p,
  cumulative = F,
  sort.prob = F,
  round = 3,
  show.label = TRUE,
  nudge = 0
)
}
\arguments{
\item{X}{Mishtane Mikri. Can be a vector of values. each X is k successes.}

\item{n}{Number of trials. If not mentioned, same as the maximum X.}

\item{cumulative}{a logical. If TRUE: draws a CDF. default is set to FALSE.}

\item{sort.prob}{a logical. Concerns the data frame returned, if TRUE: sort by probability (in decreasing order), Otherwise same order as input of X. default is set to FALSE.}

\item{round}{How many decimals to show. Default is set to 3.}

\item{show.label}{A logical. Default (TRUE) prints probability labels on the plot.}

\item{nudge}{When \code{show.label} is TRUE, you can adjust the location of the labels by inputing a 2 length vector to adjust the label position (a single values works too). Usualy no adjustment is needed.}

\item{P}{Probability for success in each trial.}
}
\value{
\code{graph} (Print only).

\code{Tohelet and Shonut} (Print only).

\code{data.frame} of X and their probs (use \code{sort.prob} to change order method).
}
\description{
By providing X, n and a probability, you can draw either CDF or PMF
}
\note{
Negative probs will be automatically corrected to absolute value, and ranged \code{0 < p < 1}.
}
\examples{
x <- c(0:5)
draw.binom(x,n=6,p=0.3)
draw.binom(x,n=6,p=0.3,cumulative=TRUE, sort.prob=TRUE)
dice <- 1:6
draw.binom(dice,p=-1/6)
## Even when using negative probability, it will work.
}
\seealso{
For a fully detailed explanation about the differences between distributions please visit \url{https://www.researchgate.net/post/What-is-the-difference-between-probability-distribution-function-and-probability-density-function}
}
