% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.exp}
\alias{draw.exp}
\title{Draws an Exponential distribution}
\usage{
draw.exp(
  X,
  rate,
  cumulative = T,
  sort.prob = F,
  round = 3,
  x.range = NULL,
  show.label = TRUE,
  nudge = 0
)
}
\arguments{
\item{X}{Mishtane Mikri. Can be a vector of values.}

\item{rate}{Rate of events (same as lambda).}

\item{cumulative}{a logical. If FALSE: draws a PMF default is set to TRUE}

\item{sort.prob}{a logical. Concerns the data frame returned, if TRUE: sort by probability (in decreasing order), Otherwise same order as input of X. default is set to FALSE.}

\item{round}{How many decimals to show. Default is set to 3.}

\item{x.range}{a vector of two numbers The probability of occurrence between these two values. Use \code{c(min,max)}.}

\item{show.label}{A logical. Default (TRUE) prints probability labels on the plot.}

\item{nudge}{When \code{show.label} is TRUE, you can adjust the location of the labels by inputing a 2 length vector to adjust the label position (a single values works too). Usualy no adjustment is needed.}
}
\value{
\code{graph} (Print only).

\code{Tohelet and Shonut} (Print only).

\code{data.frame} of X and their probs (use \code{sort.prob} to change order method, in case random numbers are assigned).

\code{range} The probability between the range in \code{x.range}. "Prob for first success between 1st and 2nd argument of the vector"
}
\description{
By providing X and a rate of events, you can draw either CDF or PMF
}
\examples{
x <- c(1:5)
draw.exp(x,rate=0.3)
draw.exp(0:5,rate=1/3,x.range=c(3,5))
## Prob for successes between this range (3rd and 5th), given P~rate of 1/3.

}
\seealso{
For a fully detailed explanation about the differences between distributions please visit \url{https://www.researchgate.net/post/What-is-the-difference-between-probability-distribution-function-and-probability-density-function}
}
