% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.geom}
\alias{draw.geom}
\title{Draws a Geometric distribution}
\usage{
draw.geom(X, p, cumulative = F, decreasing = F, round = 3)
}
\arguments{
\item{X}{Mishtane Mikri. Can be a vector of values. each X is k successes.}

\item{cumulative}{a logical. If TRUE: draws a CDF. default is set to FALSE.}

\item{decreasing}{a logical. Concerns the data frame returned, if TRUE: decreasing order of probabilities, Otherwise same order as input of X. default is set to FALSE.}

\item{round}{How many decimals to show. Default is set to 3.}

\item{P}{Probability for success in each trial.}
}
\value{
\code{graph} (Print only).

\code{Tohelet and Shonut} (Print only).

\code{data.frame} of X and their probs (use \code{decreasing} to change order method, in case random numbers are assigned).
}
\description{
By providing X and a probability, you can draw either CDF or PMF
}
\note{
Very important: Unlike usual input in \code{\link{dgeom}}, please refer to 1 as success in 1st attempt.
}
\examples{
x <- c(1:5)
draw.geom(x,p=0.3)
draw.geom(x,p=0.3,cumulative=TRUE)
}
\seealso{
For a fully detailed explanation about the differences between distributions please visit \url{https://www.researchgate.net/post/What-is-the-difference-between-probability-distribution-function-and-probability-density-function}
}
