% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.lines.R
\name{draw.lines}
\alias{draw.lines}
\title{Draw cool looking lines}
\usage{
draw.lines(
  n = 100,
  slope_range = c(0, 1),
  intercepts = 0,
  alpha_range = c(0, 0.4),
  line_color = colors(),
  line_sizes = c(0.5, 1),
  line_types = 1
)
}
\arguments{
\item{n}{The total number of lines to be drawn.}

\item{slope_range}{A vector of length two, or a single value, setting the range of slopes desired.}

\item{intercepts}{A vector of slopes (unlimited length).}

\item{alpha_range}{A vector of length two,0-1, or a single value, setting the transparency (alpha) range desired.}

\item{line_color}{A string of colors desired. Leave untouched to use all system \code{\link{colors}}.}

\item{line_sizes}{A vector of line size's desired, or a single value for a fixed size.}

\item{line_types}{Integers 0-6, single or vectorized input.}
}
\description{
Just try different values of numbers and see the different outcome. Use the example to be inspired.
}
\examples{
draw.lines()
draw.lines(line_sizes = c(0.2,1,1.5,3,4,5), # different lines sizes
slope_range = c(-2,1), # slopes ranges from -2 to 1
intercepts = c(0,1,2,3,.5), # various intercepts
alpha_range = c(0.2,.8)) # alpha ranges from 0.2 to 0.8 (by fixed jumps of 0.15)
for(. in 1:4){
          x11()
          print(draw.lines())}
}
