% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.norm}
\alias{draw.norm}
\title{Draws an Normal distribution}
\usage{
draw.norm(
  X,
  mean = 0,
  sd = 1,
  cumulative = T,
  sort.prob = F,
  round = 3,
  x.range = NULL
)
}
\arguments{
\item{X}{Mishtane Mikri. Can be a vector of values.}

\item{cumulative}{a logical. If FALSE: draws a PMF default is set to TRUE}

\item{sort.prob}{a logical. Concerns the data frame returned, if TRUE: sort by probability (in decreasing order), Otherwise same order as input of X. default is set to FALSE.}

\item{round}{How many decimals to show. Default is set to 3.}

\item{x.range}{a vector of two numbers The probability of occurrence between these two values. Use \code{c(min,max)}.}
}
\value{
\code{graph} (Print only).

\code{Tohelet and Shonut} (Print only).

\code{data.frame} of X and their probs (use \code{sort.prob} to change order method.

\code{range} The probability between the range in \code{x.range}. "Prob for success between these two values"
}
\description{
By providing X, mean and sd, you can draw either CDF or PMF
}
\examples{
x <- seq(60,90,by=5)
draw.norm(x,mean=75,sd=5)
draw.norm(x,mean=75,sd=5,cumulative=FALSE)
draw.norm(0:10,5,1.5,x.range=c(3.5,7))
# In a distribution of mean=5 and sd=1.5, print values 0-10. Also what is the probability between 3.5 and 7?
}
\seealso{
For a fully detailed explanation about the differences between distributions please visit \url{https://www.researchgate.net/post/What-is-the-difference-between-probability-distribution-function-and-probability-density-function}
}
