% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw.pois}
\alias{draw.pois}
\title{Draws a Poisson distribution}
\usage{
draw.pois(
  X,
  lam,
  cumulative = F,
  sort.prob = F,
  round = 3,
  show.label = TRUE,
  nudge = 0
)
}
\arguments{
\item{X}{Mishtane Mikri. Can be a vector of values. each X is k successes.}

\item{lam}{Rate of events.}

\item{cumulative}{a logical. If TRUE: draws a CDF. default is set to FALSE.}

\item{sort.prob}{a logical. Concerns the data frame returned, if TRUE: sort by probability (in decreasing order), Otherwise same order as input of X. default is set to FALSE.}

\item{round}{How many decimals to show. Default is set to 3.}

\item{show.label}{A logical. Default (TRUE) prints probability labels on the plot.}

\item{nudge}{When \code{show.label} is TRUE, you can adjust the location of the labels by inputing a 2 length vector to adjust the label position (a single values works too). Usualy no adjustment is needed.}
}
\value{
\code{graph} (Print only).

\code{Tohelet and Shonut} (Print only).

\code{data.frame} of X and their probs (use \code{sort.prob} to change order method, if random numbers are assigned).
}
\description{
By providing X and a rate of events, you can draw either CDF or PMF
}
\note{
Very important: Unlike usual input in \code{\link{dgeom}}, please refer to 1 as success in 1st attempt.
}
\examples{
x <- c(1:5)
draw.pois(x,lam=0.3)
draw.pois(x,lam=1.2,cumulative=TRUE)
}
\seealso{
For a fully detailed explanation about the differences between distributions please visit \url{https://www.researchgate.net/post/What-is-the-difference-between-probability-distribution-function-and-probability-density-function}
}
