% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findnorm.R
\name{find_norm}
\alias{find_norm}
\title{Find the norm of a vector}
\usage{
find_norm(vec, p = 2)
}
\arguments{
\item{vec}{A given vector.}

\item{p}{The specific dimension, default is \code{p=2}.}
}
\value{
The norm of the vector.
}
\description{
The default is finding the euclidean norm of a given vector, though you can modift the function to find any norm.
}
\examples{
find_norm(vec = c(1,2,3))
find_norm(c(0,1,0,0,1),p=0)
find_norm(c(2,-6,1.5,0,2),Inf)
}
