% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.beta.R
\name{lm.beta}
\alias{lm.beta}
\title{b to Beta Converter}
\usage{
lm.beta(m)
}
\arguments{
\item{m}{a standard object of type lm()}
}
\description{
Converts b's generated by lm() to Beta.
Results are presented in a descending order.
}
\examples{
\code{\link{lm.beta(model)}}
# model is an lm() object
}
\seealso{
\code{\link{beta2b()}} does the exact opposit.
}
