% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/median.R
\name{median}
\alias{median}
\title{Defaults median() to na.rm=TRUE}
\usage{
median(x, ..., na.rm = TRUE)
}
\description{
Simply changes the default argument to remove NA.
}
\examples{
x <- c(4,5,6,NA,6)
median(x)
## 5.5
}
