% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{mediation}
\alias{mediation}
\title{Mediation between two variables.}
\usage{
mediation(DV, x.pred, x.med, alpha = 5)
}
\arguments{
\item{DV}{The dependent variable.}

\item{x.pred}{The predictor variable.}

\item{x.med}{The mediation variable.}

\item{alpha}{Used in significant tests. default is 5 and means 0.05 .}
}
\value{
Please pay attention to the prints. They will conclude the results and warn you if there is anything to worry about.
In case something went wrong with the formula for some reason, a \code{\link{stop}} command will stop exacution.

\code{data.frame} with the results.
}
\description{
This function calculates the direct and indirect effect of two variables,
 mediated by a 3rd variable.
}
\details{
Output is a table of your 4 paths, their value and p value.
3 outputs are printed to ensure the results.
}
\note{
If some paths are significant and some aren't, check table for more details.
}
\examples{
\dontshow{attach(depression)}
mediation(DV=Bsum,x.pred = age,x.med = sat)
# In this example, we check whether satisfaction is the mediator depression score (Bsum) Height and age.
\dontrun{mediation(..., alpha =1)}
# In this example, the alpha is set to 1 (=0.01).
}
\seealso{
\code{\link{medmod}} package for more options.

\url{https://cran.r-project.org/web/packages/medmod/medmod.pdf}
}
