% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitun.R
\name{mitun}
\alias{mitun}
\title{Numerical moderation graph and summary}
\usage{
mitun(DV, x.pred, x.mod, sdH = 1, sdM = 0, sdL = 1)
}
\arguments{
\item{DV}{The dependent variable.}

\item{x.pred}{The predictor variable.}

\item{x.mod}{The moderator variable.}

\item{sdH}{Leave untouched, unless you wish to generate a different SD moderation. \code{Positive sd mod}.}

\item{sdM}{Leave untouched, unless you wish to generate a different SD moderation. \code{Mean of moderator}.}

\item{sdL}{Leave untouched, unless you wish to generate a different SD moderation.\code{Negative sd mod}.}
}
\value{
\code{\link{ggplot graph}}.

\code{\link{lm() model summary}}.

\code{\link{intercepts and slopes}} for each level of moderator.
}
\description{
This function really does it all. It will generate both a ggplot graph,
slopes&intercepts and a summary!
}
\note{
The moderator MUST be numeric.
}
\examples{
\code{\link{mitun(DV=trees$Girth,x.pred=trees$Height,x.mod=trees$Volume)}}
# This commands suggests that Girth is predicted by Height, and Volume has moderate affect on Height
mitun(..., sdH=2,sdL=2)
# In this example, the graph will generate moderations of SD=+2, mean(sdM=0), and SD=-2.
# Note that you only mention the absolute value even for the smaller SD (without - sign)
}
\seealso{
\code{\link{jtools}} or \code{\link{interactions}} for more options.
}
