% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pascal.R
\name{Pascal}
\alias{Pascal}
\title{Pascal Triangle generator}
\source{

}
\usage{
Pascal(nrow)
}
\value{
all levels up to level 30 of the triangle (or pyramid).
}
\description{
This function generates the levels of the pascal triangle.
It is highly linked with the \code{\link{choose}} function.
}
\examples{
Pascal(4)
# if you count k indexes (after 1) in row n (after 1st), you get the result of choost(n,k).
## in [[5]], which is row n=4, k=2 index is 6. this is the binomal result of choose(4,2).
read more in the URL provided.
}
\seealso{
\url{https://www.mathsisfun.com/pascals-triangle.html}
}
