% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.ci.R
\name{prop.ci}
\alias{prop.ci}
\title{Proportion Confidence Interval}
\usage{
prop.ci(n, k, conf = 95)
}
\arguments{
\item{n}{is population size.}

\item{k}{is number of success.}

\item{conf}{is confidence level desired. default at 95.}
}
\value{
\code{p} Is the succses rate, given \code{n} and \code{k}.

\code{q} Is the opposite probability of p (\code{1-p}).

\code{Lower} and \code{Upper} are the margins of the confidence interval.

\code{conf} is the confidence level.

\code{binom2norm} answers whether the assumption of normality is met.
}
\description{
Proportion Confidence Interval
}
\note{
Calculates the binomal to normal assumption: \code{p&q * n > 5}.
}
\examples{
prop.ci(n=100,k=30)
prop.ci(100,30,conf=99)
}
\seealso{
\code{\link{prop.sig}} for a proportion significance test
}
