% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prop.sig.R
\name{prop.sig}
\alias{prop.sig}
\title{Proportion Hypothesis Test}
\usage{
prop.sig(p, N, x)
}
\arguments{
\item{p}{Percents of success in population.}

\item{N}{Sample size.}

\item{x}{Percents of success in sample.}
}
\value{
\code{p} is the success rate in population.

\code{q} is the opposite probability of p (\code{1-p}).

\code{sd} is the SD estimation.

\code{Z} is the Z value, genreated given the data.

\code{binom2norm} answers whether the assumption of normality is met.
}
\description{
This function returns the Z value of the sampled proportion. Input values refer to percents.
}
\note{
Calculates the binomal to normal assumption: \code{p&q * n > 5}
}
\examples{
prop.sig(p=46,N=36,x=38)
# 46\% success in pop, while in a 36 people sample, 38\% success.
}
\seealso{
\code{\link{prop.ci}} for proportion confidence interval.
}
