% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{range}
\alias{range}
\title{Defaults range() to na.rm=TRUE}
\usage{
range(..., na.rm = TRUE)
}
\description{
Simply changes the default argument to remove NA.
}
\examples{
x <- c(4,5,6,NA,6)
range(x)
## 4 6
}
