% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression1.R
\docType{data}
\name{regression1}
\alias{regression1}
\title{A data set to practice a simple linear regression}
\format{
An object of class \code{data.frame} with 1000 rows and 2 columns.
}
\usage{
data(regression1)
}
\description{
A simple dataframe to practice simple linear regression with.
}
\references{
BGU, Psychology, Regression
}
\keyword{datasets}
