% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampdist.R
\name{sampdist}
\alias{sampdist}
\title{Sampling Distribution Generator}
\usage{
sampdist(vector, central, n, samp_size, hist = TRUE, replace = TRUE)
}
\arguments{
\item{vector}{A numeric vector.}

\item{central}{The type of central value you want to apply (mean, median, sd etc..).}

\item{n}{Number of times to sample the data from a given sample (dont confuse with sample size for each sample).}

\item{hist}{A logical. Whether to plot a histogram. Default is set to TRUE.}

\item{replace}{A logical for random sampling. Default is set to TRUE.}

\item{sam_size}{Sample size to sample each time}
}
\description{
Use this to print and plot different sample distributions
}
\details{
This function allows you the flexibility to change values (sample size, central value or number of samples) and see for yourself how the histograms and output change.
}
\examples{
times_50 <- sampdist(big5$age,mean,100,100)
# How the histogram looks when sampling 50 means?
times_5000 <- sampdist(big5$age,central=mean,n=1000,samp_size=100)
# And with 5,000?
iqrs <- sampdist(big5$age,central=IQR,n=1000,samp_size=100)
# What about the IQR? and sd? median? Try any central value with different n/samp_size.
}
\seealso{
Take a look at the well known webApp:
\url{https://onlinestatbook.com/stat_sim/sampling_dist/}.
}
