% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ttest.R
\name{se}
\alias{se}
\title{Standard Error for difference in means when SD in population is unknown}
\usage{
se(x, y)
}
\arguments{
\item{x}{A numeric vector.}

\item{y}{A second numeric vector (Independent).}
}
\description{
Calculates the standard error given two numeric vectors (may differ in length)
}
\examples{
x <- c(17,4,58,96,0,12,14)
y <- c(11,47,62,0,96,12,47,5)
se(x,y)
(t=(mean(x)-mean(y))/se(x,y))
# This is the calculation to manually find t value.
\dontrun{t.test(x,y,var.equal=TRUE)}
}
\seealso{
\url{https://www.youtube.com/watch?v=Y-bI-3kBGTY} for further explenation.
}
