% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig.z.test.R
\name{sig.z.test}
\alias{sig.z.test}
\title{Two tailed Hypothesis test when SD is known}
\usage{
sig.z.test(miu, sd, N, x, alpha = 5)
}
\arguments{
\item{miu}{Avg/mean in population.}

\item{sd}{SD in the population.}

\item{N}{Sample size.}

\item{x}{Mean found in sample.}

\item{alpha}{Alpha decided. default 5.}
}
\value{
\code{H0} \code{rejected} or \code{accepted}.

\code{Z} the Z value.

\code{Z crit} Criteria according to the \code{alpha} provided.

\code{alpha} in both sides (takes the \code{alpha} input and divides by 2).
}
\description{
The test is two tailed for when the SD in population is known.
# p value for two tails. default Z crit is +-1.96.
}
\examples{
sig.z.test(620,80,25,588)
# In this example, 620 is the known mean with sd of 80. A 25 size sample produced a mean of 588. Is it significant?
sig.z.test(miu=620,sd=80,N=25,x=588,alpha=1)
# Now we changed the alpha from 0.05 to 0.01.
}
