% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmodel.R
\name{ssmodel}
\alias{ssmodel}
\title{Generates the SST/Res/Reg, Rsq and Adj Rsq of lm model.}
\usage{
ssmodel(model)
}
\description{
Make sure to create your \code{lm} object first. Put your object in the function and see how magic happens.
}
\examples{
x <- lm(Volume~.,trees)
ssmodel(x)
}
\seealso{
What happens if you use the mean in your model as prediction?
\code{ssmodel(lm(trees$Volume~rep(mean(trees$Volume),nrow(trees))))}
# This is a hard-to-read input that contains a repetition vector of the mean value of Volume for as many times as the  number of rows in the data,
# then these values (actually, constant) are used to predict Volume. See for yourself that SSRes=SST and SSReg=0.
}
