% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use.packages.R
\name{use.package}
\alias{use.package}
\title{Install and/or call multiple packages}
\usage{
use.package(packages, library.only = TRUE)
}
\arguments{
\item{packages}{A string of 1 or more packages names to be installed. Be aware of typos.}

\item{library.only}{A logical. When TRUE, function only calls required packages, otherwise, packages will be installed. Default is set to TRUE.}
}
\description{
A vectorized version of \code{\link{library}} or \code{\link{install.packages}} that allows you to install and call from memory multiple packages at once.
}
\note{
If you ask to re-install an existing package, a restart might be necessary. just press YES when asked by R. Otherwise abort and set \code{library.only} to TRUE.
}
\examples{
use.package("tidyverse")
my_pack <- c("ggplot2","magrittr")
use.package(my_pack,library.only=FALSE) # This will re-install these packages, Abort and set library only to TRUE.
}
