% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikicloud.R
\name{wikicloud}
\alias{wikicloud}
\title{A wordcloud generator from Wikipedia values}
\usage{
wikicloud(
  page,
  language = "en",
  bg_color = "white",
  word_color = c("#76acdc", "#4c4c4c"),
  shape = "circle",
  min_freq = 0,
  max_freq = NULL,
  min_length = 1,
  max_length = 25,
  remove_pattern = "[:punct:]",
  remove_words = NULL
)
}
\arguments{
\item{page}{The name of the page as it appears in Wikipedia. Pay attention to Capita letters and undeline (incase of multiple words in a single value).}

\item{language}{The language of Wikipedia page you wish. default is set to English ("en"). type the short language symbol \code{"en","fr","es"} and etc..}

\item{bg_color}{The background color. default is white. a single string with quotes: \code{"white"}.}

\item{word_color}{The color of the words pressented. This is a un-limited vector in size. Use basic colors in R such as \code{c("red","steelblue","deeppink","brown")} and/or HTML codes such as \code{"#76acdc"}.}

\item{shape}{Choose your desired cloud shape. \code{circle} (default), \code{cardioid, diamond, triangle, triangle-forward, pentagon, star}.}

\item{min_freq}{Minimum frequency of the word to display. Default is 1 (Once).}

\item{max_freq}{Maximum frequency of the word to display. Default is the maximum frequency.}

\item{min_length}{Minimum word's length to display. Default is 1.}

\item{max_length}{Maximum word's length to display. Default is 25.}

\item{remove_pattern}{A pattern to ignore, such as \code{[:punct:]}. It's best to leave untouched.}

\item{remove_words}{Words to ignore. Function is set to ignore repetitive words such as \code{a, and, the} and etc.. If you do change this one, use free text, such as \code{"just type any words in one sentence and R will seperate them and ignore each"}.}
}
\value{
\code{\link{data.frame}} with the unique words and their frequency.
}
\description{
By typing any Wikipedia value you will get a beautiful wordcloud, with word's frequency as its size. The function is highly modifiable.
}
\details{
This is a combination of Data mining (using \code{\link{rvest}}), Data cleaning (using the \code{\link{tidyverse}}) and word cloud visualization, using \code{\link{wordcloud2}}.
Just type your desired Wikipedia value and run, or if you wish you can modify many things, such as color, frequancy, langauge and more..
}
\examples{
#Simpelest way, just plot a word cloud, using default settings:
wikicloud("RStudio")
wikicloud("RStudio",shape="star",min_freq=2)
# Changing the shape of the cloud and filtering words that appear only once.

wikicloud("Tag cloud",shape="triangle",bg_color="blue",word_color=c("red","deeppink","#418122"))
# This is how the value" Tag cloud" is on Wikipedia's URL.

wikicloud("Donald Trump",language="es",word_color="orange")
# Changing the language of the Wiki page.

wikicloud("donald trump",min_freq=5,max_freq=60)
# By selecting the min and max frequency, you can modify the range of frequency. you can set either. both, or none (for full word table, leave untouched).
}
\seealso{
\url{https://yannapps.shinyapps.io/Wikipedia_Word_Cloud/}
}
