% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta2b.R
\name{beta2b}
\alias{beta2b}
\title{Convert Beta to b}
\usage{
beta2b(beta, x, y)
}
\arguments{
\item{beta}{The beta value you want to convert.}

\item{x}{The x variable (Undependent).}

\item{y}{The y variable (Dependent).}
}
\value{
raw Mekadem b
}
\description{
Converts the Beta value to a raw b, given the x and y
}
\examples{
beta2b(beta=1.45, x=trees$Height, y=trees$Volume)
## 3.740674
## For every increase of 1 unit in Height, Volume increases by 3.740674 units.
}
