% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.binom.R
\name{draw.binom}
\alias{draw.binom}
\title{Draws a Binomal distribution, Tohelet and Shonut}
\usage{
draw.binom(X, n, p, round = 3)
}
\arguments{
\item{X}{Possible overall succsess in all trials (Mishtane Mikri).}

\item{n}{Number of trials.}

\item{p}{The probability for success in each trial.}

\item{round}{Number of decimals desired. default is set to 3.}
}
\value{
\code{\link{data.frame}} with the specific X and its probability.
  \code{\link{print}}: two verbal prints: Tohelet and Shonut of your experiment.
}
\description{
By giving the desired X (Mishtane Mikri), number of trials and succsses rate,
You will get a clean graph, plus a \code{\link{data.frame}} and two prints of Tohelet and Shonut.
}
\details{
There are 2 red balls and 3 white ones in a bucket. Each time you pick a ball,
you return it to the bucket (replace=TRUE).
Succsess is to pull a red ball. You have n=3 trials. What are your X and their probability?
\code{draw.binom(X=c(0,1,2,3),n=3,2/5)}.
}
\note{
when using large numbers, \code{\link{ggplot2}} graph might go off. Sorry about that. See data instead.
}
\examples{
draw.binom(10:12,12,0.85) # 10,11 or 12 succsses in 12 trials with 85\% chance for succsess.

You can change the decimals by adjusting the 'round' argument.
 draw.binom(c(0,1,2),2,3/5,round=5)
}
