% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draws a  distribution, Tohelet and Shonut}
\usage{
draw(X, P, round = 3)
}
\arguments{
\item{X}{Mishtane Mikri.}

\item{round}{Number of decimals desired. default is set to 3.}

\item{p}{The probability for each x 'P(x)'}
}
\value{
\code{\link{data.frame}} with the specific X and its probability.
  \code{\link{print}}: two verbal prints: Tohelet and Shonut of your experiment.
}
\description{
Generates a graph, Tohelet and Shonut, plus a \code{\link{data.frame}} with your data.
}
\details{
By supplying the X and P(x), This function will generate the Tohelet and Shonut, and a clean graph.
}
\examples{
X=10:12
probs=c(0.4,0.25,0.35)
draw(X,probs)
# Distribution of 10, 11 & 12, with their probs as created in 'X' and 'Probs'.

# You can change the decimals by adjusting the 'round' argument.
 draw(c(0,1,2),c(0.2,0.5,0.3),round=5)
}
\seealso{
Use \code{\link{draw.binom}} to draw a binomal distribution.
}
