% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.beta.R
\name{lm.beta}
\alias{lm.beta}
\title{b to Beta Converter}
\usage{
lm.beta(m)
}
\arguments{
\item{m}{a standard object of type lm()}
}
\value{
A data frame of the betas in an absolute decending order.
}
\description{
This function converts b's generated by lm() to Beta.
Results are presented in a descending order.
}
\examples{
model= lm(Bsum~age+sat,depression)
lm.beta(model)
}
\seealso{
\code{\link{beta2b()}} converts Beta back to raw B.
}
