% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.beta.R
\name{lm.beta}
\alias{lm.beta}
\title{b to Beta Converter}
\usage{
lm.beta(lm.model)
}
\arguments{
\item{lm.model}{a standard object of type \code{lm}}
}
\value{
A data frame of the betas in an absolute descending order.
}
\description{
This function converts b's generated by \code{\link{lm}} to Beta.
Results are presented in a descending order.
}
\examples{
model= lm(Bsum~age+sat,depression)
lm.beta(model)
Gen_model=lm(Generosity~.,happiness)
# Generosity predicted by all other variables
lm.beta(Gen_model)
}
\seealso{
\code{\link{beta2b}} converts Beta back to raw B.
}
