% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{mediation}
\alias{mediation}
\title{Mediation between two variables.}
\usage{
mediation(DV, x.pred, x.med, alpha = 5)
}
\arguments{
\item{DV}{The dependent variable.}

\item{x.pred}{The predictor variable.}

\item{x.med}{The mediation variable.}

\item{alpha}{Used in significant tests. default is 5 and means 0.05 .}
}
\value{
[1] \code{\link{a*b+c'=c}} formula is tested.

[2] Paths's significance.

[3] Is there a direct effect of \code{\link{c'}}.

[4] \code{\link{data.frame}} with the paths (values and p value).
}
\description{
This function calculates the direct and indirect effect of two variables,
 mediated by a 3rd variable.
}
\details{
Output is a table of your 4 paths, their value and p value.
3 outputs are printed to ensure the results.
}
\note{
If some paths are significant and some aren't, check table for more details.
}
\examples{
mediation(DV=trees$Girth,x.pred=trees$Height,x.med=trees$Volume)
# In this example, we check whether volume is the mediator between Height and Girth.
mediation(..., alpha =1)
# In this example, the alpha is set to 1, or 0.01.
}
\seealso{
\code{\link{medmod}} package for more options.
}
