% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mitun.R
\name{mitun}
\alias{mitun}
\title{Numerical moderation graph and summary}
\usage{
mitun(DV, x.pred, x.mod, sdH = 1, sdM = 0, sdL = -1)
}
\arguments{
\item{DV}{The dependent variable.}

\item{x.pred}{The predictor variable.}

\item{x.mod}{The moderator variable.}

\item{sdH}{Leave untouched, unless you wish to generate a different SD moderation. \code{Positive sd mod}.}

\item{sdM}{Leave untouched, unless you wish to generate a different SD moderation. \code{Mean of moderator}.}

\item{sdL}{Leave untouched, unless you wish to generate a different SD moderation.\code{Negative sd mod}.}

\item{sdH/M/L}{refer to SD level of the moderator. feel free to explore, but default is best for analsys.}
}
\value{
\code{\link{ggplot}} graph.

\code{\link{lm}} model summary.

\code{intercepts and slopes} for each level of moderator.
}
\description{
This function really does it all. It will generate both a ggplot graph,
slopes&intercepts and a summary!
}
\note{
The moderator MUST be numeric. This function works for a simple  moderation model.
}
\examples{
mitun(DV=data.moderation$depression,
          x.pred=data.moderation$suicide_tendencies,
          x.mod=data.moderation$optimism)
# This commands suggests that depression is predicted by suicide tendencies, and optimism has moderate affect on depression
mitun(..., sdH=2,sdL=-2)
# In this example, the graph will generate moderation of SD=+2, mean(sdM=0), and SD=-2 in the moderator variable.
}
\seealso{
\code{\link{jtools}} or \code{\link{interactions}} for more options.
}
