% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mode.R
\name{mode}
\alias{mode}
\title{Mode finder}
\usage{
mode(x, freq = F, show.na = F)
}
\arguments{
\item{x}{A numeric vector}

\item{freq}{A logical argument. Adds the frequency of your mode, default is FALSE}

\item{show.na}{A logical argument. Counts NA as a value.}
}
\value{
Most frequent value, and if freq is set to TRUE, how many times it appears (including NAs).
}
\description{
What value is the most frequent one?
}
\details{
In case of two Mo values, both will return.
}
\examples{
x <- c(4,3,2,3)
mode(x)
y <- c(x,2)
mode(y)
# Set frequancy to TRUE:
mode(y,freq=TRUE)
# Set show.na to TRUE
z <- c(4,3,2,3,2,NA,NA)
mode(z,show.na=TRUE)
mode(z,T,T)
}
