% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semi.partial.cor.R
\name{semi.partial.cor}
\alias{semi.partial.cor}
\title{Partial and Semi-partial correlations}
\usage{
semi.partial.cor(DV, x.predictor, x.controlled)
}
\value{
\code{data.frame} with the \code{Rsq} of partial and semi-partial, sample size and DF.
}
\description{
In the Partial correlation, the controlled variable is removed from both DV and UDV,
While in Semi-partial, controlled is removed only from the UDV, and so it is very much alike the Beta (direct effect).
}
\details{
Be sure to understand the theoretical meaning of semi and partial correlation.
In a nutshell, both correlate residuals created by a third variable.
 Partial correlation is the correlation of two
 variables while controlling a third.
  SP holds constant the controlled variable only for the UDV (predictor).
}
\note{
Partial will always be Greater than/Equal to Semi-partial.
}
\examples{
\dontshow{attach(depression)}
semi.partial.cor(DV=Bsum,x.predictor=m_years, x.controlled=age)
# We correlate 'Bsum' and 'm_years', holding 'age' constant. In SP it's been removed from only the predicting variable,
# whilst in partial from both.
}
\seealso{
\code{\link{ppcor}} for more options.
}
