% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wikicloud.R
\name{wikicloud}
\alias{wikicloud}
\title{A wordcloud generator from Wikipedia values}
\usage{
wikicloud(
  page,
  language = "en",
  bg_color = "white",
  word_color = c("#76acdc", "#4c4c4c"),
  shape = "circle",
  min_freq = 0,
  remove_pattern = "[.?:;,-]",
  remove_words = NULL
)
}
\arguments{
\item{page}{The name of the page as it appears in Wikipedia. Pay attention to Capita letters and undeline (incase of multiple words in a single value).}

\item{language}{The language of Wikipedia page you wish. default is set to English ("en"). type the short language symbol \code{"en","fr","es"} and etc..}

\item{bg_color}{The background color. default is white. a single string with quotes: \code{"white"}.}

\item{word_color}{The color of the words pressented. This is a un-limited vector in size. Use basic colors in R such as \code{c("red","steelblue","deeppink","brown")} and/or HTML codes such as \code{"#76acdc"}.}

\item{shape}{Choose your desired cloud shape. \code{ ‘circle’ (default), ‘cardioid’, ‘diamond’ (alias of square), ‘triangle-forward’, ‘triangle’, ‘pentagon’, and ‘star’}.}

\item{min_freq}{Minimum frequency of the word in order to appear in the cloud and table. Default is 1 (that is, every word).}

\item{remove_pattern}{A pattern to ignore, such as \code{.?:;-}. Its best to leace untouched.}

\item{remove_words}{Words to ignore. If left as NULL, R will automatically ignore non-meaningful words in English, Spanish and French. It is advised to leave untouchd. If you do change this one, use free text, such as \code{"just type any words in one sentence and R will seperate them and ignore each"}.}
}
\value{
\code{\link{data.frame}} with the unique words and their frequency.
}
\description{
By typing any Wikipedia value you will get a beautiful wordcloud, with word's frequency as its size. The function is highly modifiable.
}
\details{
This is a combination of Data mining (using \code{\link{rvest}}), data cleaning (using multiple packages, but mostly \code{\link{stringr}}) and a beautiful string visualization, using \code{\link{wordcloud2}}.
You dont need to change anything to make it work, and if you wish you can modify many things, from the color of the words and background, to the shape of the cloud, minimal frequency, language of Wiki value and more!
}
\note{
It is advised to simply install \code{\link{tidyverse}}. It includes many useful packages.
}
\examples{
#Simpelest way, just plot a word cloud, using default settings:
wikicloud("RStudio")
wikicloud("RStudio",shape="star",min_freq=2)
# Changing the shape of the cloud and filtering words that appear only once.
wikicloud("Tag_cloud",shape="triangle",bg_color="blue",word_color=c("red","deeppink","#418122"))
# This is how the value" Tag cloud" is on Wikipedia's URL.
wikicloud("Trump",language="es",word_color="orange")
# Changing the language of the Wiki page.
## You might want to remove words yourself in this one, by modifying the 'remove_words' argument.
}
